refactor(backend): consolidate error-message helper into a single util
`toErrorMessage` was redefined inline (identical body) in 5 files across
the application and presentation layers, while a richer `getErrorMessage`
already lived in infrastructure/utils.

- Move `isError`/`getErrorMessage` to application/utils/error.utils.ts, a
  location all layers can import (infrastructure depends inward).
- Delete the 5 inline `toErrorMessage` definitions; route every call
  through `getErrorMessage`.
- Update the infrastructure imports to the new path; remove the old
  infrastructure/utils/error.utils.ts.

Part of #152.
