+++ apps/backend/src/application/utils/error.utils.ts
