+++ apps/backend/src/infrastructure/external/spotify-artist.client.ts
@@ -1,5 +1,5 @@
 import type { SettingsPort } from "@/application/ports/settings.port";
-import { getErrorMessage } from "../utils/error.utils";
+import { getErrorMessage } from "../../application/utils/error.utils";
 import { CircuitBreaker } from "./circuit-breaker";
 import { PromiseCache } from "./promise-cache";
 import { RateLimiter } from "./rate-limiter";
