+++ apps/backend/src/infrastructure/external/spotify-playlist.client.ts
@@ -2,7 +2,7 @@ import { NormalizedTrack } from "@spotiarr/shared";
 import type { SettingsPort } from "@/application/ports/settings.port";
 import { AppError } from "@/domain/errors/app-error";
 import { SpotifyUrlHelper } from "@/domain/helpers/spotify-url.helper";
-import { getErrorMessage } from "../utils/error.utils";
+import { getErrorMessage } from "../../application/utils/error.utils";
 import { CircuitBreaker } from "./circuit-breaker";
 import { PromiseCache } from "./promise-cache";
 import { RateLimiter } from "./rate-limiter";
