+++ apps/backend/src/presentation/controllers/events.controller.ts
@@ -1,9 +1,6 @@
 import { Request, Response } from "express";
+import { getErrorMessage } from "../../application/utils/error.utils";
 import { resolveAllowedOrigin } from "../middleware/cors";
-
-function toErrorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
-}
 
 interface SseClient {
   id: number;
@@ -54,7 +51,7 @@ export class EventsController {
       try {
         client.res.write(payload);
       } catch (error) {
-        console.error("Failed to write SSE event:", toErrorMessage(error));
+        console.error("Failed to write SSE event:", getErrorMessage(error));
       }
     }
   };
